Committing latest changes as of 01/22/2020.
* Refactored code

* Removed obsolete properties
 - Changed all references of Spooler.Magic to Spooler.Context
 - Updated SpoolerMagicConverter to support SpoolerContext type

* Better type and usage of Spooler.Context property

* Removed unnecessary Spooler.StrLen property

* Changed IDetailProvider implementation methods into extensions

* Other various changes

+++ Antilli/Source/MainWindow.xaml.cs
@@ -342,7 +342,7 @@ namespace Antilli
             {
             case ".dam":
                 modelFile.SpoolerLoaded += (s, e) => {
-                    if (s.Context == (int)ChunkType.CharacterSkeletons)
+                    if (s.Context == ChunkType.CharacterSkeletons)
                     {
                         var skel = s as SpoolableBuffer;
